Oficina/requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.85.1
fastapi-utils==0.2.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
pydantic==1.10.2
pyserial==3.5
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
